+++ Dependencies/infoware/premake5.lua
@@ -1,0 +1,76 @@
+project "infoware"
+	kind "StaticLib"
+    language "C++"
+    cppdialect "C++17"
+	staticruntime "on"
+	
+	targetdir ("Bin/" .. outputdir .. "/%{prj.name}")
+	objdir ("Bin-Intermediate/" .. outputdir .. "/%{prj.name}")
+
+	warnings "Off"
+
+    includedirs 
+    {
+     	"include/",  
+    }
+
+    files
+    {
+    	"src/**.cpp",
+    }
+    excludes
+    {
+    	"src/pci.cpp"
+    }
+
+    -- check https://github.com/ThePhD/infoware/blob/main/CMakeLists.txt for compilation settings
+    defines	
+    {
+    	"INFOWARE_VERSION=\"0.6.0\"",
+   		"INFOWARE_USE_OPENGL=1",
+    }
+
+	--- Platform ---
+	filter "system:windows"
+		systemversion "latest"
+		defines
+		{
+		}
+		links
+		{
+		}
+
+
+	filter "system:linux"
+		defines
+		{
+		}
+		links
+		{
+		}
+
+	filter "system:macosx"
+		defines
+		{
+		}	
+		links
+		{
+		}
+
+
+	--- Configurations ---
+	filter "configurations:Debug"
+		runtime "Debug"
+		symbols "Full" --vs17 and newer | symbols "On"
+		optimize "Off" -- No optimization will be performed.
+
+	filter "configurations:Release"
+		runtime "Release"
+		optimize "On" -- Perform a balanced set of optimizations.
+		inlining "Explicit" -- Only inline functions explicitly marked with the inline keyword.
+
+
+	filter "configurations:Dist"
+		runtime "Release"
+		optimize "Full" -- Full optimization.
+		inlining "Auto" -- Inline any suitable function for full performance